Docker/todo-app/backend/requirements.txt
flask==3.0.3
gunicorn==21.2.0
psycopg2-binary== 2.9.11